fix several small functions to 100% match
- BufStreamNAND::Eof: swap mSize==mTell to mTell==mSize (load order)
- BufStreamNAND::DeleteChecksum: remove null guard, use delete directly
- FileCacheFile::Eof: swap mParent->Size()<=mPos to mPos>=mParent->Size()
- UsbWii::SetInactive: cast num to unsigned int for cmplwi
- HttpWii::CancelAsync: compare NHTTPCancelConnection result ==0 instead of bool cast
- BeatMap::Interpolate(int): hoist BeatInfo refs before k1 for r30/r31 allocation

+++ src/system/usbwii/UsbWii.cpp
@@ -192,7 +192,7 @@ void UsbWii::SetLED(int num, int led) {
 }
 
 void UsbWii::SetInactive(int num) {
-    if (num >= 4)
+    if ((unsigned int)num >= 4)
         return;
     sDevices[num].flags = (sDevices[num].flags & ~kUsbFlagActive) | kUsbFlagInactive;
 }
